fix integer
