+++ ft_integer/ft_abs.c
@@ -1,20 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_min.c                                           :+:      :+:    :+:   */
+/*   ft_abs.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/30 19:53:04 by reasuke           #+#    #+#             */
-/*   Updated: 2023/11/30 19:55:02 by reasuke          ###   ########.fr       */
+/*   Created: 2024/02/09 13:17:40 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:15:17 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-int	ft_min(int a, int b)
+int	ft_abs(int a)
 {
-	if (b < a)
-		return (b);
+	if (a < 0)
+		return (-a);
 	return (a);
 }
