+++ ft_integer/ft_min.c
@@ -1,22 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_swap.c                                          :+:      :+:    :+:   */
+/*   ft_min.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/30 20:29:07 by reasuke           #+#    #+#             */
-/*   Updated: 2023/11/30 20:38:51 by reasuke          ###   ########.fr       */
+/*   Created: 2023/11/30 19:53:04 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:15:43 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-void	ft_swap(int *a, int *b)
+int	ft_min(int a, int b)
 {
-	int	tmp;
-
-	tmp = *a;
-	*a = *b;
-	*b = tmp;
+	if (b < a)
+		return (b);
+	return (a);
 }
