+++ ft_integer/ft_swap.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_max.c                                           :+:      :+:    :+:   */
+/*   ft_swap.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/30 19:51:36 by reasuke           #+#    #+#             */
-/*   Updated: 2023/11/30 19:52:01 by reasuke          ###   ########.fr       */
+/*   Created: 2023/11/30 20:29:07 by reasuke           #+#    #+#             */
+/*   Updated: 2024/02/24 12:15:49 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
-int	ft_max(int a, int b)
+void	ft_swap(int *a, int *b)
 {
-	if (a < b)
-		return (b);
-	return (a);
+	int	tmp;
+
+	tmp = *a;
+	*a = *b;
+	*b = tmp;
 }
