+++ ft_printf/print_string/print_str.c
@@ -6,11 +6,12 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 17:02:59 by reasuke           #+#    #+#             */
-/*   Updated: 2024/02/23 23:49:17 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/02/24 12:12:08 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../ft_printf.h"
+#include "ft_integer.h"
 
 void	fp_print_str(char *str, t_format_info *fi, t_format_result *fr, int fd)
 {
